functions/github-view-counter/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
appwrite>=2.0.0
python-dotenv>=0.19.0
Pillow>=8.3.1 